[LINALG] Add value tensor variant to `fill_.Scalar`
This commit adds the op `PseudoAtenFillScalarOp` that represents
`AtenFill_ScalarOp` without the underscore. The approach is the same
as in commit dd998fa4d4163af14519fed436f29e82e72673ae.

Adding this op allows for a simpler and more consistent version of the
`empty` and `empty_like` op e2e tests.