linalg: add support for lowering AtenMeanDimOp
This patch augments the existing logic for lowering reduction operations
to also support lowering of AtenMeanDimOp operations.  The key idea
behind lowering mean operations is to generate _two_ linalg operations,
one for computing the sum reduction on the specified dimensions and
another for performing an elementwise multiplication with the reciprocal
of the size of the tensor along the reduced dimensions.

This patch also adds unit and end-to-end tests for validating the
lowering of this operation.