linalg: support more cases in lowering reshape op
Prior to this patch, PyTorch's reshape and view operations were lowered
to either `tensor.expand_shape` or `tensor.collapse_shape` ops, which
prevented us from lowering reshape operations for which (1) the source
and destination shape involved both collapsing as well as expansion or
(2) the source and destination shapes had the same rank but not the same
values in each dimension.

A better solution is to use `tensor.reshape`, which handles both of the
above-mentioned cases.  This patch augments the existing lowering of
reshape operations from the torch dialect to the tensor dialect, and it
adds two end-to-end tests to validate the translation.

However, upstream LLVM currently supports lowering reshape operations
only when all shape values are statically known, so this patch adds a
guard condition to ensure that the subsequent lowering to the LLVM
dialect will not fail.