simplify-shape-calc: propagate sizes across copy statements
The `CopyToValueTensorOp` and `CopyToNonValueTensorOp` statements don't
change the size of the tensor.  By making the shape refinement passes
aware of this, the compiler now produces fewer unranked tensors, which
results in fewer translation failures downstream.  Specifically, this
patch looks for so-called downcast operations, where the tensor size is
dropped from the source tensor to the result, before relaying the size
of the source tensor to the uses of the downcast op.

The drawback of this change is that it introduces additional
`TensorStaticInfoCastOp` ops, but these ops should become NOPs later in
the pass pipeline.

To prevent a significant increase in compilation times, this patch
limits the shape propagation to at most six uses of the downcast op.