build: streamline options for enabling LTC and MHLO
Prior to this patch, LTC was enabled by default, but it was disabled in
setup.py and in the Github CI rules.  MHLO, on the other hand, was
enabled by default, but also (i.e. redundantly) enabled in setup.py and
in the Github CI rules.

This had the effect of LTC being left enabled unless the build is
triggered by setup.py or Github CI.  It also had the effect of forcing
us to update the option in multiple places, whenever it needed to be
changed.

This patch sets the options' default value at the option definintion,
and it explicitly sets the value only when the default is overriden,
like for disabling MHLO in in-tree builds on macOS.